requirements.txt
aiogram>=3.7.0
python-dotenv>=1.0.1
yt-dlp>=2025.1.0
google-api-python-client>=2.155.0
google-auth>=2.37.0
google-auth-oauthlib>=1.2.1
